[staking] Fix for build, move transaction under staking/types
