+++ node/worker/worker_test.go
@@ -5,18 +5,16 @@ import (
 	"math/rand"
 	"testing"
 
-	types2 "github.com/harmony-one/harmony/staking"
-
-	blockfactory "github.com/harmony-one/harmony/block/factory"
-	chain2 "github.com/harmony-one/harmony/internal/chain"
-
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethdb"
+	blockfactory "github.com/harmony-one/harmony/block/factory"
 	"github.com/harmony-one/harmony/common/denominations"
 	"github.com/harmony-one/harmony/core"
 	"github.com/harmony-one/harmony/core/types"
 	"github.com/harmony-one/harmony/core/vm"
+	chain2 "github.com/harmony-one/harmony/internal/chain"
 	"github.com/harmony-one/harmony/internal/params"
+	types2 "github.com/harmony-one/harmony/staking/types"
 )
 
 var (
