+++ staking/types/staking_transaction.go
